bot/requirements.txt
aiohttp>=3.9
twitchio>=2.9,<3
python-dotenv>=1.0
